refactor: centralize env config and domain types
