+++ components/steps/Step4Permits/ConfinedSpace/SafetyManager.tsx
@@ -3,27 +3,10 @@
 
 import { create } from 'zustand';
 import { persist } from 'zustand/middleware';
-import { createClient } from '@supabase/supabase-js';
+import { supabase } from '@/lib/supabase';
 
-// =================== CONFIGURATION SUPABASE ROBUSTE ===================
-const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL || 'https://your-project.supabase.co';
-const supabaseKey = process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY || 'your-anon-key';
-
-let supabase: any = null;
-let supabaseEnabled = false;
-
-try {
-  if (supabaseUrl && supabaseKey && supabaseUrl !== 'https://your-project.supabase.co') {
-    supabase = createClient(supabaseUrl, supabaseKey);
-    supabaseEnabled = true;
-    console.log('✅ SafetyManager: Supabase configuré');
-  } else {
-    console.log('📝 SafetyManager: Utilisation du localStorage');
-  }
-} catch (error) {
-  console.log('⚠️ SafetyManager: Supabase non configuré, utilisation du localStorage');
-  supabaseEnabled = false;
-}
+// =================== CONFIGURATION SUPABASE ===================
+const supabaseEnabled = true;
 
 // =================== TYPES UNIVERSELS ===================
 export type ProvinceCode = 'QC' | 'ON' | 'BC' | 'AB' | 'SK' | 'MB' | 'NB' | 'NS' | 'PE' | 'NL';
