+++ lib/config.ts
@@ -1,0 +1,10 @@
+export function getEnvVar(key: string): string {
+  const value = process.env[key];
+  if (!value) {
+    throw new Error(`Missing environment variable: ${key}`);
+  }
+  return value;
+}
+
+export const SUPABASE_URL = getEnvVar('NEXT_PUBLIC_SUPABASE_URL');
+export const SUPABASE_ANON_KEY = getEnvVar('NEXT_PUBLIC_SUPABASE_ANON_KEY');
